pyproject.toml
[project]
name = "avito-pr-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Vladislav",email = "vlad.karmanchikofff@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.121.2,<0.122.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "alembic (>=1.17.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.5,<0.15.0)",
    "loguru (>=0.7.3,<0.8.0)"
]

[tool.ruff]
line-length = 120
target-version = "py312"
fix = true
exclude = [
    ".git",
    ".github",
    ".idea",
    ".venv",
    "__pycache__",
    "alembic",
    "settings",
    "tests",
]

[tool.ruff.lint]
select = [
    "E",            # pycodestyle errors (основной стандарт PEP8)
    "F",            # pyflakes (логические ошибки и неиспользуемое)
    "B",            # bugbear (опасные паттерны)
    "I",            # isort (сортировка импортов)
    "UP",           # pyupgrade (совместимость с последней версией)
    "S",            # bandit (простые проверки безопасности)
    "ANN",          # type annotations (аннотации типов)
    "SIM",          # simplify (упрощения)
    "T20",          # print отладочный
    "ARG",          # неиспользуемые аргументы
    "PL",           # pylint-подобные правила
]
ignore = [
    "ANN001",       # не требовать type hints для self
    "S101",         # разрешаем assert проверки session в репозиториях
    "B008",         # разрешаем вызывать методы внутри Depends в ручках
    "PLR0913",      # выключаем too many arguments (> 5)
    "PLR6301",      # выключаем method could be static
    "PLR0904",     # выключаем too many public methods
]

[tool.ruff.lint.isort]
default-section = "third-party"
combine-as-imports = true
known-first-party = ["infrastructure", "app", "settings"]
section-order = ["standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.format]
preview = false
quote-style = "single"
indent-style = "space"